# PERF: Performance Improvement on `DataFrame.to_csv()` when `index=False`