@mroeschke 2024-08-26
For you benchmark could you show before and after timings