@KevsterAmp 2024-08-26
Output before performance improvement
```
Operation A time: 869.2354643344879 seconds
Operation B time: 42.1906418800354 seconds
```

Output (after performance improvement)
```
Operation A time: 51.408071756362915 seconds
Operation B time: 45.78637385368347 seconds
```

**Operation B** is used for time comparison when resetting index, the change improves the performance on **Operation A**

### Added output times to the description as well